BcAppAPI/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
bson==0.5.10
cffi==1.14.5
click==8.0.1
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
fastapi-users==6.1.0
h11==0.12.0
idna==3.2
loguru==0.5.3
makefun==1.11.3
motor==2.4.0
passlib==1.7.4
Pillow==8.2.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
python-multipart==0.0.5
qrcode==6.1
redis==3.5.3
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
Werkzeug==2.0.1
wrapt==1.12.1
